+++ sample-app/src/components/CometChatHome/CometChatHome.tsx
@@ -25,7 +25,7 @@ import { CometChatUserDetails } from "../CometChatDetails/CometChatUserDetails";
 import { CometChatThreadedMessages } from "../CometChatDetails/CometChatThreadedMessages";
 import { CometChatCallDetails } from "../CometChatCallLog/CometChatCallLogDetails";
 import { CometChatAlertPopup } from "../CometChatAlertPopup/CometChatAlertPopup";
-import { CometChatAvatar,CometChatButton,CometChatConfirmDialog, CometChatConversationEvents,CometChatGroupEvents,CometChatGroupMembers,CometChatGroups,CometChatIncomingCall,CometChatMessageEvents,CometChatToast,CometChatUIKit,CometChatUIKitConstants,CometChatUIKitLoginListener,CometChatUIKitUtility,CometChatUserEvents,CometChatUsers,CometChatUIEvents,localize,IMessages,IMouseEvent,IActiveChatChanged,MessageStatus,IGroupMemberAdded,IGroupMemberKickedBanned} from "@cometchat/chat-uikit-react";
+import { CometChatAvatar,CometChatButton,CometChatConfirmDialog, CometChatConversationEvents,CometChatGroupEvents,CometChatGroupMembers,CometChatGroups,CometChatIncomingCall,CometChatMessageEvents,CometChatToast,CometChatUIKit,CometChatUIKitConstants,CometChatUIKitLoginListener,CometChatUIKitUtility,CometChatUserEvents,CometChatUsers,CometChatUIEvents,localize,IMessages,IMouseEvent,IActiveChatChanged,MessageStatus,IGroupMemberAdded,IGroupMemberKickedBanned, IGroupMemberJoined} from "@cometchat/chat-uikit-react";
 
 import { CallLog } from "@cometchat/calls-sdk-javascript";
 
@@ -286,7 +286,7 @@ function CometChatHome(props: { theme?: string }) {
         const updateisFirstChat = ({ message, status }: IMessages) => {
             const receiverId = message.getReceiverId();
             const sender = message.getSender();
-            if ((appState.selectedItemUser && (appState.selectedItemUser.getUid() == receiverId || (appState.selectedItemUser.getUid() == sender.getUid() && receiverId == loggedInUser?.getUid())) || appState.selectedItemGroup && (appState.selectedItemGroup.getGuid() == receiverId || loggedInUser?.getUid() == receiverId)) && isFreshChatRef.current && status == MessageStatus.success) {
+            if ((appState.selectedItemUser && (appState.selectedItemUser.getUid() == receiverId || ((!sender || (sender && appState.selectedItemUser.getUid() == sender.getUid())) && receiverId == loggedInUser?.getUid())) || appState.selectedItemGroup && (appState.selectedItemGroup.getGuid() == receiverId || loggedInUser?.getUid() == receiverId)) && isFreshChatRef.current && status == MessageStatus.success) {
                 setAppState({ type: 'updateIsFreshChat', payload: false });
                 isFreshChatRef.current = false;
                 let conversationWith = appState.selectedItemUser ? appState.selectedItemUser.getUid() : appState.selectedItemGroup?.getGuid();
@@ -1075,6 +1075,7 @@ function CometChatHome(props: { theme?: string }) {
                                     if (loggedInUser) {
                                         const groupClone = CometChatUIKitUtility.clone(group);
                                         groupClone.setHasJoined(false);
+                                        groupClone.setMembersCount(groupClone.getMembersCount() - 1);
                                         CometChatGroupEvents.ccGroupLeft.next({
                                             userLeft: CometChatUIKitUtility.clone(loggedInUser),
                                             leftGroup: groupClone,
@@ -1363,6 +1364,11 @@ function CometChatHome(props: { theme?: string }) {
         const ccOpenChat = CometChatUIEvents.ccOpenChat.subscribe((item) => {
             openChatForUser(item.user);
         })
+        const ccGroupJoineed = CometChatGroupEvents.ccGroupMemberJoined.subscribe((data: IGroupMemberJoined) => {
+            setGroup(data.joinedGroup)
+            setSelectedItem(data.joinedGroup);
+            setAppState({ type: "updateSelectedItemGroup", payload: data.joinedGroup });
+        })
 
         const ccClickEvent = CometChatUIEvents.ccMouseEvent.subscribe((mouseevent: IMouseEvent) => {
             if (mouseevent.event.type === "click" && (mouseevent.body as { CometChatUserGroupMembersObject: User })?.CometChatUserGroupMembersObject) {
@@ -1401,6 +1407,7 @@ function CometChatHome(props: { theme?: string }) {
             ccConversationDeleted?.unsubscribe();
             ccOpenChat?.unsubscribe();
             ccClickEvent?.unsubscribe();
+            ccGroupJoineed?.unsubscribe();
         };
     }, [newChat, selectedItem]);