+++ sample-app/src/metaInfo.ts
@@ -1,6 +1,6 @@
 export const metaInfo = {
   name: "sample-app",
-  version: "5.0.6",
+  version: "5.0.7",
   type: "sample",
   platform: "React",
 };