+++ src/components/Calling/CallingExtensionDecorator.tsx
@@ -262,7 +262,7 @@ export class CallingExtensionDecorator extends DataSourceDecorator {
       callBuilder.setCallListener(
         new CometChatUIKitCalls.OngoingCallListener({
           onCallEndButtonPressed: () => {
-            CometChatUIEvents.ccShowOngoingCall.next({ child: null });
+            CometChatUIEvents.ccShowOngoingCall.next({ child: null,message });
           },
         })
       );
@@ -271,7 +271,7 @@ export class CallingExtensionDecorator extends DataSourceDecorator {
     const ongoingCallScreen = (
       <CometChatOngoingCall callSettingsBuilder={callBuilder} sessionID={sessionId} callWorkflow={CallWorkflow.directCalling} />
     );
-    CometChatUIEvents.ccShowOngoingCall.next({ child: ongoingCallScreen });
+    CometChatUIEvents.ccShowOngoingCall.next({ child: ongoingCallScreen,message });
   }