+++ src/components/Calling/CometChatCallButtons/CometChatCallButtons.tsx
@@ -9,7 +9,7 @@ import { CometChatSoundManager } from "../../../resources/CometChatSoundManager/
 import { CometChatUIKitUtility } from "../../../CometChatUIKit/CometChatUIKitUtility";
 import { localize } from "../../../resources/CometChatLocalize/cometchat-localize";
 import { CometChatUIKitConstants } from "../../../constants/CometChatUIKitConstants";
-import { MessageStatus } from "../../../Enums/Enums";
+import { CallWorkflow, MessageStatus } from "../../../Enums/Enums";
 import { CometChatButton } from "../../BaseComponents/CometChatButton/CometChatButton";
 import audioCall from "../../../assets/audio_call_button.svg";
 import videoCall from "../../../assets/video_call_button.svg";
@@ -467,6 +467,7 @@ const CometChatCallButtons = (props: CallButtonsPropsType) => {
         <CometChatOngoingCall
           sessionID={sessionIdRef.current}
           callSettingsBuilder={getCallBuilder()}
+          callWorkflow={activeGroup ? CallWorkflow.directCalling : CallWorkflow.defaultCalling}
         />
       ) : null}