+++ src/utils/DataSourceDecorator.ts
@@ -86,13 +86,13 @@ export abstract class DataSourceDecorator implements DataSource {
   getStatusInfoView(
     message: CometChat.BaseMessage,
     alignment: MessageBubbleAlignment,
-    hideReceipt?: boolean,
+    hideReceipts?: boolean,
     datePattern?: DatePatterns
   ) {
     return (this.dataSource ?? new MessagesDataSource()).getStatusInfoView(
       message,
       alignment,
-      hideReceipt,
+      hideReceipts,
       datePattern
     );
   }